Move ReadAndProcessOnTheFly from cp2k.py to engineparts.py
